Trigger WelcomeEmail Notification
- Once user successfully created an account, trigger welcome email to user's email.
