+++ public/views/index.html
@@ -1,5 +1,6 @@
-<script src="js/templates/vendor.min.js"></script>
+
 <script src="js/templates/scripts.min.js"></script>
+<script src="js/templates/vendor.min.js"></script>
 
 <!-- Page Content-->
 <!-- Main Slider-->
